backend/requirements.txt
Flask
Flask-CORS
Flask-JWT-Extended
Flask-SQLAlchemy
PyMySQL
cryptography
python-dotenv
